refactor: Moves Basic classes declarations to Basics.h (GDMBasics.h now only includes all required files)
