+++ lib/GDMBasics/src/Game.cpp
@@ -2,7 +2,7 @@
 // Created by elly_sparky on 23/01/24.
 //
 
-#include "GDMBasics.h"
+#include "Basics.h"
 
 namespace mate
 {
