+++ lib/GDMBasics/src/Trigger.cpp
@@ -1,7 +1,7 @@
 //
 // Created by elly_sparky on 14/08/24.
 //
-#include "GDMBasics.h"
+#include "Basics.h"
 
 namespace mate
 {
